# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Social Media Platform API
  description: |
    A comprehensive API for the AI-powered social media management platform.
    
    This API provides endpoints for user management, content creation, social media integration,
    AI-powered features, and analytics.
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API requests are rate limited to prevent abuse:
    - 100 requests per 15 minutes per IP address
    - Premium users have higher limits
    
    ## Error Handling
    
    The API returns standard HTTP status codes and JSON error responses:
    ```json
    {
      "message": "Error description",
      "error": "Detailed error information (development only)"
    }
    ```
  version: 1.0.0
  contact:
    name: AI Social Media Team
    email: support@aisocialmedia.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.aisocialmedia.com
    description: Production server
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      description: Create a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                company:
                  type: string
                  example: Acme Corp
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and return JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      summary: User logout
      description: Logout user (client-side token removal)
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
...
  /api/content:
    get:
      summary: Get user's content
      description: Retrieve all content for the authenticated user with pagination
      tags:
        - Content
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by content status
          schema:
            type: string
            enum: [draft, scheduled, published, archived]
        - name: platform
          in: query
          description: Filter by platform
          schema:
            type: string
            enum: [facebook, instagram, twitter, linkedin, tiktok, youtube]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Content retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Content'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create new content
      description: Create a new piece of content
      tags:
        - Content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentInput'
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  content:
                    $ref: '#/components/schemas/Content'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (development only)
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
          description: JWT authentication token
        user:
          $ref: '#/components/schemas/UserBasic'
    UserBasic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
        subscription:
          type: string
          enum: [free, basic, premium, enterprise]
...
    Content:
      type: object
      properties:
        id:
          type: string
        user:
          type: string
        title:
          type: string
        content:
          type: object
          properties:
            text:
              type: string
            hashtags:
              type: array
              items:
                type: string
            mentions:
              type: array
              items:
                type: string
        media:
          type: array
          items:
            $ref: '#/components/schemas/Media'
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/PlatformContent'
        postType:
          type: string
          enum: [text, image, video, carousel, story, reel, poll]
        status:
          type: string
          enum: [draft, scheduled, published, archived]
        aiGenerated:
          $ref: '#/components/schemas/AIGenerated'
        analytics:
          $ref: '#/components/schemas/Analytics'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContentInput:
      type: object
      required:
        - title
        - content
        - platforms
      properties:
        title:
          type: string
        content:
          type: object
          required:
            - text
          properties:
            text:
              type: string
            hashtags:
              type: array
              items:
                type: string
            mentions:
              type: array
              items:
                type: string
        media:
          type: array
          items:
            $ref: '#/components/schemas/Media'
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/PlatformContent'
        postType:
          type: string
          enum: [text, image, video, carousel, story, reel, poll]
          default: text
        tags:
          type: array
          items:
            type: string
        campaign:
          type: string
        scheduling:
          type: object
          properties:
            timezone:
              type: string
            optimalTime:
              type: boolean
    Media:
      type: object
      properties:
        type:
          type: string
          enum: [image, video, gif, document]
        url:
          type: string
        filename:
          type: string
        size:
          type: number
        dimensions:
          type: object
          properties:
            width:
              type: number
            height:
              type: number
        duration:
          type: number
        thumbnail:
          type: string
    PlatformContent:
      type: object
      properties:
        platform:
          type: string
          enum: [facebook, instagram, twitter, linkedin, tiktok, youtube]
        postId:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published, failed]
        scheduledFor:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        platformSpecific:
          type: object
          properties:
            text:
              type: string
            hashtags:
              type: array
              items:
                type: string
            mentions:
              type: array
              items:
                type: string
        error:
          type: string
    AIGenerated:
      type: object
      properties:
        isAIGenerated:
          type: boolean
        prompt:
          type: string
        model:
          type: string
        confidence:
          type: number
          format: float
        suggestions:
          type: array
          items:
            type: string
    Analytics:
      type: object
      properties:
        views:
          type: number
        likes:
          type: number
        shares:
          type: number
        comments:
          type: number
        clicks:
          type: number
        reach:
          type: number
        impressions:
          type: number
        engagement:
          type: object
          properties:
            rate:
              type: number
              format: float
            score:
              type: number
              format: float
    Pagination:
      type: object
      properties:
        current:
          type: integer
        pages:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...